+++ src/components/ListItem/ListItem.stories.js
@@ -24,9 +24,7 @@ storiesOf("ListItem", module)
           unit: "kg",
           step: 0.1,
           price: 35000,
-          category: "bumbu",
-          promo: "Diskon",
-          promo_desc: "Diskon Akhir Pekan"
+          category: "bumbu"
         })}
         count={number("Count", 1)}
         quantity={text("Quantity", "100 gram")}
